requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
filelock==3.8.0
huggingface-hub==0.10.1
idna==3.4
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
schedule==1.1.0
tokenizers==0.13.2
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12
